u32 for GC marker
This aligns with the C definition. Memory is not reduced because of the
Rust 8 byte alignment.

+++ src/redisearch_rs/c_entrypoint/inverted_index_ffi/src/lib.rs
@@ -529,7 +529,7 @@ pub unsafe extern "C" fn InvertedIndex_LastId(ii: *const InvertedIndex) -> t_doc
 /// The following invariant must be upheld when calling this function:
 /// - `ii` must be a valid, non NULL, pointer to an `InvertedIndex` instance.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn InvertedIndex_GcMarker(ii: *const InvertedIndex) -> usize {
+pub unsafe extern "C" fn InvertedIndex_GcMarker(ii: *const InvertedIndex) -> u32 {
     debug_assert!(!ii.is_null(), "ii must not be null");
 
     // SAFETY: The caller must ensure that `ii` is a valid pointer to an `InvertedIndex`
